feat-petvoter

+++ Pith/src/Controller/Api/PetController.php
@@ -77,6 +77,7 @@ class PetController extends AbstractController
      */
     public function edit(Pet $pet, Request $request)
     {
+        $this->denyAccessUnlessGranted('PET_EDIT', $pet);
         $form = $this->createForm(PetType::class, $pet, ['csrf_protection' => false]);
 
         $json = $request->getContent();
@@ -104,6 +105,8 @@ class PetController extends AbstractController
      */
     public function delete(Pet $pet)
     {
+        $this->denyAccessUnlessGranted('PET_DELETE', $pet);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($pet);
         $em->flush();
